Make the 3 Trits singletons in the module.
Since there are only 3 possible values for a Trit, and they are
immutable, it doesn't make sense to construct multiple copies of them.
It also doesn't make sense to lookup the integer value on demand, when
we could just store the relevant integer in each of the 3 instances.

Also, modify the tests to mostly check against the string value of the
trits, rather than the integer value (why compare 4 bytes when you could
compare 1?).