Add sequence support and tests for Trits objects.
Next up will be the tritwise logic and comparison operators, also
concatenation (+) and repetition (*).  Then on to the integer
arithmetic.

Still not totally sure whether I want to implement the integer stuff in
a subclass or just allow Trits objects to be used as integers.  One
advantage to the subclass option is that we can use + to mean
concatention with Trits, but addition with Int.