Add integer division by subtraction to Int.
This division algorithm seems to get the right results (test still needs
to be written) but it going to be awfully slow as the operands get large.

After I get the tests written it might be worth trying to improve on
the performance.

I also need to make sure that the results are returned as Int objects --
at the moment some cases lead to a Trits object.