Add a basic instruction set for the T3.
I still need to write proper tests for this, but informal testing has
produced encouraging results.

I'd also like to add a debugging/verbose mode to the T3 which outputs
each instruction as it is executed.  It might make sense to add a 'name'
attribute to Instruction so that they can be more easily identified.