Add a verbose mode to the T3 for easier tracing and debugging.
In verbose mode, the T3 emits a line to stdout for every instruction it
executes, showing the opcode, instruction name and operand.

This involved added an optional 'name' attribute to the Instruction
class.  As a side benefit, it also makes the instruction table for the
T3 a bit easier to understand.