Add logical instructions to T3: AND, OR, XOR and NOT.
Also change the shift instructions to write their results into the
default register, rather than shifting the contents of the operand
register in-place.  There are arguments in favour of both behaviours,
but putting the result in DR is what all the other operations do, so I
think POLA recommends doing the same with shifting.